requirements.txt
numpy==1.21.2
adversarial-robustness-toolbox==1.8.0
pillow==8.2.0
scipy==1.7.1
pyyaml==5.4.1
torch
torchvision
torchaudio
matplotlib==3.5.0